% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_bad_equations.R
\name{add_biomass}
\alias{add_biomass}
\title{Add column \code{biomass}, by evaluating \code{equation_allomerty} for each \code{dbh}.}
\usage{
add_biomass(.data)
}
\arguments{
\item{.data}{A dataframe with columns \code{equation_allometry} and \code{dbh}.}
}
\value{
A dataframe.
}
\description{
Add column \code{biomass}, by evaluating \code{equation_allomerty} for each \code{dbh}.
}
\examples{
equations <- pick_equations(allodb::scbi_species, "species")

allodb::scbi_tree1 \%>\%
  add_equation(equations) \%>\%
  # Just a few for speed
  head() \%>\%
  add_biomass()
}
