% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_bad_equations.R
\name{add_equation}
\alias{add_equation}
\title{Add the column \code{equation_allometry} to a census dataframe.}
\usage{
add_equation(census, equations, .f = dplyr::inner_join)
}
\arguments{
\item{census}{A ForestGEO-like census dataframe, with columns \code{sp} \code{dbh}.}

\item{equations}{A dataframe with columns \code{sp} and \code{equation_allometry}.}

\item{.f}{A \code{dplyr::*_join()} function -- e.g. \code{\link[dplyr:left_join]{dplyr::left_join()}}
\code{\link[dplyr:full_join]{dplyr::full_join()}} \code{\link[dplyr:inner_join]{dplyr::inner_join()}} (default).}
}
\value{
A dataframe.
}
\description{
Add the column \code{equation_allometry} to a census dataframe.
}
\examples{
equations <- pick_equations(allodb::scbi_species, "species")
add_equation(census = allodb::scbi_tree1, equations)
}
