% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGB_computation.R
\name{AGB.comp}
\alias{AGB.comp}
\title{AGB computation}
\usage{
AGB.comp(site, D, WD, H = NULL)
}
\arguments{
\item{site}{provide the full name of your site (in lower case) i.e. 'barro colorado island'}

\item{D}{a column with tree diameters (in mm)}

\item{WD}{a column with wood density values}

\item{H}{a column with tree heights (optional)}
}
\value{
a vector with AGB values (in Mg)
}
\description{
Compute above-ground biomass using the updated model of Chave et al. (2014), given in Rejou-Mechain et al. (2017)
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
