% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allo_temperate}
\alias{allo_temperate}
\title{Allometric equations form temperate sites.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 415 rows and 25 columns.}
\source{
Erika B. Gonzalez (\link{GonzalezEB@si.edu}).
}
\usage{
allo_temperate
}
\description{
Allometric equations form temperate sites.
}
\keyword{datasets}
