% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allodb-package.r
\docType{package}
\name{allodb}
\alias{allodb}
\alias{allodb-package}
\alias{allodb-package}
\title{`allodb`: A database of allometric equations for ForestGeo.}
\description{
`allodb`: A database of allometric equations for ForestGeo.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
