% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmss_default_eqn.R
\name{allodb_eqn_crucial}
\alias{allodb_eqn_crucial}
\title{Crucial columns form \strong{allodb} equations-table used by \code{as_default_eqn()}.}
\usage{
allodb_eqn_crucial()
}
\value{
A string.
}
\description{
Crucial columns form \strong{allodb} equations-table used by \code{as_default_eqn()}.
}
\examples{
allodb_eqn_crucial()
}
