% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{allodb_example}
\alias{allodb_example}
\title{Help create path to example data.}
\usage{
allodb_example(path)
}
\arguments{
\item{path}{Directory containing example.}
}
\value{
A path.
}
\description{
Help create path to example data.
}
\examples{
allodb_example("csv_database")
}
