% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_allodb_master.R
\name{as_allodb}
\alias{as_allodb}
\title{Create objects of class allodb.}
\usage{
as_allodb(x)
}
\arguments{
\item{x}{A dataframe from \strong{allodb}.}
}
\value{
A dataframe subclass "allodb".
}
\description{
This function creates S3 objects of class "allodb". Compared with the data
directly delivered in the \strong{allodb} package, objects of class "allodb" have
two main differences:
\itemize{
\item The column types are not all text but the types that is most useful for
computations.
\item Missing values are represented in a single way: with \code{NA}.
}
}
\details{
The datasets directly delivered by \strong{allodb} are optimized for delivering
the greatest amount of information in the most robust and flexible way. All
values are stored as text. This allows, for example, different to store
different representation of missing values. The trade-off is that text values
don't allow direct computation. First  they may need to be converted to other
types of data, such as integers.
}
\examples{
as_allodb(allodb::equations)
}
\seealso{
Other functions to interact with the database: \code{\link{master}}
}
\concept{functions to interact with the database}
