% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_winner_loser.R
\name{bind_winner_loser}
\alias{bind_winner_loser}
\title{Keep rows from a dataframe and add missing rows from another dataframe.}
\usage{
bind_winner_loser(winner, loser)
}
\arguments{
\item{winner}{Dataframes which rows to always keep.}

\item{loser}{Dataframe from which to add missing rows to the winner
dataframe.}
}
\value{
A dataframe.
}
\description{
Keep rows from a dataframe and add missing rows from another dataframe.
}
\examples{
winner <- tibble(rowid = 1, x = "from winner")
loser <- tibble(rowid = 1:2, x = "from loser")
bind_winner_loser(winner, loser)

winner2 <- tibble(rowid = 1:2, x = "from winner")
loser2 <- tibble(rowid = 1:2, x = "from loser")
bind_winner_loser(winner2, loser2)
}
