% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmss.R
\name{bmss}
\alias{bmss}
\title{Calculate biomass.}
\usage{
bmss(dbh_sp, default_eqn)
}
\arguments{
\item{dbh_sp}{A dataframe created with \code{\link[=bmss_cns]{bmss_cns()}}.}

\item{default_eqn}{A dataframe created with \code{\link[=bmss_default_eqn]{bmss_default_eqn()}}.}
}
\value{
A dataframe.
}
\description{
Calculate biomass.
}
\examples{
dbh_sp <- bmss_cns(scbi_tree1, scbi_species, site = "scbi")
default_eqn <- bmss_default_eqn(allodb::master())
bmss(dbh_sp, default_eqn)
}
