% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmss_default_eqn.R
\name{bmss_cns}
\alias{bmss_cns}
\title{Create a census table of the kind useful in \strong{bmss}.}
\usage{
bmss_cns(census, species, site)
}
\arguments{
\item{census}{A ForestGEO-like census-dataframe.}

\item{species}{A ForestGEO-like species-dataframe.}

\item{site}{The name of the site. One of \code{allodb::sites_info$site}.}
}
\value{
A dataframe.
}
\description{
Create a census table of the kind useful in \strong{bmss}.
}
\examples{
bmss_cns(allodb::scbi_tree1, allodb::scbi_species, site = "scbi")
}
