% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmss_default_eqn.R
\name{bmss_default_eqn}
\alias{bmss_default_eqn}
\title{Structure equaitons from \strong{allodb} as a \strong{bmss} default-equations table.}
\usage{
bmss_default_eqn(.data)
}
\arguments{
\item{.data}{An object of class 'allodb_eqn' built with \code{\link[allodb:master]{allodb::master()}}.}
}
\value{
A dataframe.
}
\description{
Structure equaitons from \strong{allodb} as a \strong{bmss} default-equations table.
}
\examples{
bmss_default_eqn()
}
