% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_equations.R
\name{choose_equation}
\alias{choose_equation}
\title{Function to add new equations to the equation table (then used in the
get_biomass function)}
\usage{
choose_equation(taxa = "all", climate = "all", new_equations = NULL)
}
\arguments{
\item{taxa}{character vector with taxa to be kept. Default is \code{all}, in which
case all taxa are kept.}

\item{climate}{character vector with name of Koppen climate to be kept.
Default is \code{all}, in which case all climates are kept.}

\item{new_equations}{optional equation table, as created by the
\code{add_equations} function. Default is \code{NULL}, in which case the default
equation table is used.}
}
\value{
A subsetted equation table.
}
\description{
Function to add new equations to the equation table (then used in the
get_biomass function)
}
\examples{

}
