% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_equations.R
\name{choose_equations}
\alias{choose_equations}
\title{Subset the default equation table, based on taxa, koppen climate zone or region.}
\usage{
choose_equations(
  taxa = "all",
  climate = "all",
  region = "all",
  new_equations = NULL
)
}
\arguments{
\item{taxa}{character vector with taxa to be kept. Default is \code{all}, in which
case all taxa are kept.}

\item{climate}{character vector with name of Koppen climate to be kept.
Default is \code{all}, in which case all climates are kept.}

\item{region}{character vector with name of location(s) or country(ies) or broad region(s)
(\code{Europe}, \verb{North America}, etc -> add complete list) to be kept.
Default is \code{all}, in which case all regions are kept.}

\item{new_equations}{optional equation table, as created by the
\code{add_equations} function. Default is \code{NULL}, in which case the default
equation table is used.}
}
\value{
A subsetted equation table.
}
\description{
The subsetted equation table can then be used in the \code{get_biomass} function.
}
\examples{

}
