% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_eqn.R
\name{combine_eqn}
\alias{combine_eqn}
\title{Combine multiple tables of allometric equations.}
\usage{
combine_eqn(...)
}
\arguments{
\item{...}{One or multiple tables of allometric equations.}
}
\value{
A dataframe.
}
\description{
Combine multiple tables of allometric equations.
}
\examples{
x <- tibble::tribble(~sp, ~equation, ~id, ~dbh,
                   "sp1",        NA, 200,   33,
                   "sp2", "3 * dbh", 111,   40,
                   "sp3", "7 * dbh", 221,   15,
                   "sp4", "5 * dbh", 332,   40)
x

stem <- table_eqn(x = x, eqn = "equation", sp = "sp", dbh = "dbh", site = "bci",
  eqn_type = "stem")

sp <- table_eqn(x = x, eqn = "equation", sp = "sp", dbh = "dbh", site = "bci",
  eqn_type = "species")

site <- table_eqn(x = x, eqn = "equation", sp = "sp", dbh = "dbh", site = "bci",
  eqn_type = "site")

tibble_all_eqn(stem, site, sp)
}
\seealso{
table_equations
}
