% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{csv_master}
\alias{csv_master}
\title{Read csv files into a master table.}
\usage{
csv_master(path)
}
\arguments{
\item{path}{Path to directory with equations.csv, sitespecies.csv, and
sites_info.csv.}
}
\value{
A dataframe.
}
\description{
Read csv files into a master table.
}
\examples{
csv_master(allodb_example("csv_database"))
# Or
\dontrun{
csv_master("data-raw/csv_database")
}
}
