% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{List datasets in a package.}
\usage{
datasets(package)
}
\arguments{
\item{package}{String giving the name of a package.}
}
\value{
A character vector.
}
\description{
List datasets in a package.
}
\examples{
datasets("allodb")
}
