% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_eqn.R
\docType{data}
\name{default_eqn}
\alias{default_eqn}
\title{Default equations.}
\format{An object of class \code{default_eqn} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 619 rows and 6 columns.}
\usage{
default_eqn
}
\description{
Default equations.
}
\keyword{datasets}
