% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGB_computation.R
\name{density.ind}
\alias{density.ind}
\title{Assign wood density}
\usage{
\method{density}{ind}(df, site, wsgdata, denscol = "wsg")
}
\arguments{
\item{df}{a data.table}

\item{site}{provide the full name of your site (in lower case) i.e. 'barro colorado island'}

\item{wsgdata}{a list of tree species (mnenomic) and corresponding wood density}

\item{denscol}{the variable to be return ("wsg" by default). No other option is implemented.}
}
\value{
a data.table (data.frame) with all relevant variables.
}
\description{
Assign wood density using CTFS wood density data base (WSG)
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
