% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_bad_equations.R
\name{drop_bad_equations}
\alias{drop_bad_equations}
\title{Drop rows known to contain problematic equations.}
\usage{
drop_bad_equations(.data)
}
\arguments{
\item{.data}{An \strong{allodb} dataframe with the variable \code{equation_id}.}
}
\value{
A dataframe.
}
\description{
Drop rows known to contain problematic equations.
}
\examples{
drop_bad_equations(master())
}
\keyword{internal}
