% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations}
\alias{equations}
\alias{equations_metadata}
\title{Tables of allometric equations and associated metadata for ForestGEO sites.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 178 rows and 24 columns.}
\usage{
equations

equations_metadata
}
\description{
A dataset of best available allometry equations to calculate AGB per species
per ForestGEO site:
\itemize{
\item \code{equations}: Table of allometric equations.
\item \code{equations_metadata}: Table of metadata of \code{equations}.
}
}
\keyword{datasets}
