% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_params.R
\name{est_params}
\alias{est_params}
\title{Calibrate new allometric equations}
\usage{
est_params(
  genus,
  coords,
  species = NULL,
  new_eqtable = NULL,
  wna = 0.1,
  w95 = 500,
  nres = 10000
)
}
\arguments{
\item{genus}{a character vector, containing the genus (e.g. "Quercus") of
each tree.}

\item{coords}{a numeric vector of length 2 with longitude and latitude (if
all trees were measured in the same location) or a matrix with 2 numerical
columns giving the coordinates of each tree.}

\item{species}{a character vector (same length as genus), containing the
species (e.g. "rubra")  of each tree. Default is "NULL", when no species
identification is available.}

\item{new_eqtable}{Optional. An equation table created with the
\code{new_equations()} function. Default is the compiled \pkg{allodb}  equation table.}

\item{wna}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine
the dbh-related and sample-size related weights attributed to equations
without a specified dbh range or sample size, respectively. Default is 0.1}

\item{w95}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine
the value at which the sample-size-related weight reaches 95\% of its
maximum value (max=1). Default is 500.}

\item{nres}{number of resampled values. Default is "1e4".}
}
\value{
A data frame (\code{tibble::tibble()} object) of fitted coefficients (columns) of the non-linear
least-square regression \deqn{AGB = a * dbh ^ b + e,  with e ~ N(0, sigma^2)}
}
\description{
This function calibrates new allometric equations from sampling previous
ones. New allometric equations are calibrated for each species and location
by resampling the original compiled equations; equations with a larger sample
size, and/or higher taxonomic rank, and climatic similarity with the species
and location in question are given a higher weight in this process.
}
\examples{
# calibrate new allometries for all Lauraceae species
lauraceae <- subset(scbi_stem1, Family == "Lauraceae")
est_params(
  genus = lauraceae$genus,
  species = lauraceae$species,
  coords = c(-78.2, 38.9)
)
}
\seealso{
\code{\link[=weight_allom]{weight_allom()}}, \code{\link[=new_equations]{new_equations()}}
}
