% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicated_rowid.R
\name{evaluate_equations}
\alias{evaluate_equations}
\alias{find_duplicated_rowid}
\alias{pick_one_row_by_rowid}
\title{Evaluate equations, find duplicated ones, and pick a single row by \code{rowid}.}
\usage{
evaluate_equations(.data)

find_duplicated_rowid(.data)

pick_one_row_by_rowid(.data)
}
\arguments{
\item{.data}{A dataframe as those created with \code{\link[=pick_best_equations]{pick_best_equations()}}.}
}
\value{
A dataframe with a single row by each value of \code{rowid}.
}
\description{
Evaluate equations, find duplicated ones, and pick a single row by \code{rowid}.
}
\examples{
library(dplyr)

best <- allodb::scbi_tree1 \%>\%
  census_species(allodb::scbi_species, "scbi") \%>\%
  get_equations() \%>\%
  pick_best_equations()

# Has duplicated rowid
find_duplicated_rowid(best)

# NO duplicated rowid
pick_one_row_by_rowid(best)

find_duplicated_rowid(pick_one_row_by_rowid(best))
}
