% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ficus}
\alias{ficus}
\title{Data to from tropical forests.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67 rows and 8 columns.}
\source{
Ervan Rutishauser (\link{er.rutishauser@gmail.com}), received on Sat,
Oct 28, 2017 at 4:05 AM.
}
\usage{
ficus
}
\description{
Data to from tropical forests.
}
\seealso{
Other data from tropical forests: \code{\link{WSG}},
  \code{\link{site.info}}, \code{\link{wsg_ctfs3}}
}
\keyword{datasets}
