% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_default_eqn.R
\name{find_duplicated_eqn}
\alias{find_duplicated_eqn}
\title{Find duplicated equations.}
\usage{
find_duplicated_eqn(.data = allodb::default_eqn)
}
\arguments{
\item{.data}{A dataframe -- usually of subclass "default_eqn".}
}
\value{
A dataframe.
}
\description{
Find duplicated equations.
}
\examples{
find_duplicated_eqn()
}
