% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/genus_family.R
\docType{data}
\name{genus_family}
\alias{genus_family}
\title{Title}
\format{
An object of class \code{data.frame} with 248 rows and 2 columns.

An object of class \code{data.frame} with 248 rows and 2 columns.
}
\source{
?
}
\usage{
genus_family

genus_family
}
\description{
Description

Description ...
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{missing_values_metadata}},
\code{\link{references_metadata}},
\code{\link{sites_info}},
\code{\link{sitespecies}},
\code{\link{wsg}}
}
\concept{database datasets}
\keyword{datasets}
