% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass.R
\name{get_biomass}
\alias{get_biomass}
\title{Function to compute aboveground biomass (or other variables ) from allometric
equations.}
\usage{
get_biomass(
  dbh,
  h = NULL,
  genus = rep(NA, length(dbh)),
  species = NULL,
  coords,
  var = c("Total aboveground biomass", "Whole tree (above stump)"),
  add_weight = FALSE,
  use_height_allom = TRUE
)
}
\arguments{
\item{dbh}{A numerical vector containing tree diameter at breast height (dbh)
measurements, in cm}

\item{h}{A numerical vector (same length as dbh) containing tree height
measurements, in m. Default is NULL, when no measurement is available.}

\item{genus}{A character vector (same length as dbh), containing the genus
(e.g. "Quercus") of each tree. Default is NULL, when no identification is
available.}

\item{species}{A character vector (same length as dbh), containing the
species (e.g. "rubra")  of each tree. Default is NULL, when no
identification is available.}

\item{coords}{A numerical vector of length 2 with longitude and latitude (if
all trees were measured in the same location) or a matrix with 2 numerical
columns giving the coordinates of each tree. Default is NULL when no
information is available.}

\item{var}{What dependent variable(s) should be provided in the output? Default
is \verb{Total aboveground biomass} and \verb{Whole tree (above stump)}, other possible values are:
\verb{Bark biomass}, \code{Branches (dead)}, \code{Branches (live)}, \verb{Branches total (live, dead)},
\verb{Foliage total}, \code{Height}, \code{Leaves}, \verb{Stem (wood only)}, \verb{Stem biomass},
\verb{Stem biomass (with bark)}, \verb{Stem biomass (without bark)}, \verb{Whole tree (above and belowground)}.
Be aware that only a few equations exist for those other variables, so estimated values might not be
very acurate.}

\item{add_weight}{Should the relative weigth given to each equation in the
\code{equations} data frame be added to the output? Default is FALSE.}

\item{use_height_allom}{A logical value: should the height allometries from
Bohn et al (2014) be used in the AGB allometries from Jansen et al (1996)?
Default is TRUE.}
}
\value{
A vector of class "numeric" of the same length as dbh, containing AGB
value (in kg) for every stem, or the dependent variable as defined in
\code{var}.
}
\description{
This function creates S3 objects of class "numeric".
}
\examples{
data(scbi_stem1)
get_biomass(
dbh = scbi_stem1$dbh,
genus = scbi_stem1$genus,
species = scbi_stem1$species,
coords = c(-78.2, 38.9)
)
}
