% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eqn.R
\name{get_eqn}
\alias{get_eqn}
\title{Get default equations of each type.}
\usage{
get_eqn(dbh_species)
}
\arguments{
\item{dbh_species}{A dataframe as those created with \code{\link[=bmss_cns]{bmss_cns()}}.}
}
\value{
A nested dataframe with each row containing the data of an equation
type.
}
\description{
Get default equations of each type.
}
\examples{
dbh_species <- bmss_cns(dbh_sp, species, site)
get_eqn(dbh_species)
}
