% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equations.R
\name{get_equations}
\alias{get_equations}
\title{Get default equations of each type.}
\usage{
get_equations(dbh_species)
}
\arguments{
\item{dbh_species}{A dataframe as those created with \code{\link[=census_species]{census_species()}}.}
}
\value{
A nested dataframe with each row containing the data of an equation
type.
}
\description{
Get default equations of each type.
}
\examples{
dbh_species <- census_species(allodb::scbi_tree1, allodb::scbi_species, "scbi")
get_equations(dbh_species)
}
