% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gymno_genus}
\alias{gymno_genus}
\title{Gymnosperms identified in selected ForestGEO sites}
\format{
A data frame with 95 observations and 3 variables: ```
[,1]	Family	character [,2]	Genus	character [,3] conifer logical\preformatted{
[,1]: R:,1
[,2]: R:,2
[,3]: R:,3
}
}
\usage{
gymno_genus
}
\description{
Table with genus and their associated family for Gymnosperms identified in the
ForestGEO sites used in allodb. This data frame is particularly important to
differentiate conifers as input in the weight_allom() function.
}
\details{
\itemize{
\item \code{gymno_genus}: A data frame with genus and families for Gymnosperms
identified in the extratropical ForestGEO sites used in allodb.
}
}
\keyword{datasets}
