% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_function.R
\name{make_function}
\alias{make_function}
\title{Make a function.}
\usage{
make_function(args, body, env = parent.frame())
}
\arguments{
\item{args}{Arguments.}

\item{body}{Body.}

\item{env}{Environment.}
}
\value{
A function.
}
\description{
Make a function.
}
\examples{
args <- alist(x = x, a = 1, b = NULL)
body <- quote(a * x + b)
myfun <- make_function(args, body)

myfun

myfun(x = 2, a = 4, b = 2)
}
\author{
joran (profile: https://goo.gl/6biFxw;
contribution: https://goo.gl/Bg7p5K).
}
