% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{master}
\alias{master}
\title{Join the main allodb tables: \code{equations``, }sitespecies\code{, and }sites_info`.}
\usage{
master()
}
\arguments{
\item{.f}{A function, form the \strong{dplyr} package, used to join tables (e.g.
\code{dplyr::full_join()}).}
}
\value{
A dataframe.
}
\description{
Join the main allodb tables: \code{equations``, }sitespecies\code{, and }sites_info`.
}
\examples{
master()

\dontrun{
# Nice view in RStudio
View(master())
}
}
\seealso{
\code{\link[dplyr:full_join]{dplyr::full_join()}}.

Other functions to interact with the database: \code{\link{set_type}}
}
\concept{functions to interact with the database}
