% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_values_metadata}
\alias{missing_values_metadata}
\title{Table of metadata associated to the table \code{missing_values}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 7 columns.}
\usage{
missing_values_metadata
}
\description{
Table of metadata associated to the table \code{missing_values}.
}
\keyword{datasets}
