% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowbind_inorder.R
\name{pick_best_equations}
\alias{pick_best_equations}
\title{Find the best equations given a priority order.}
\usage{
pick_best_equations(.data, order = c("species", "genus", "family",
  "mixed_hardwood", "woody_species"))
}
\arguments{
\item{.data}{List of dataframes. Each should have a \code{rowid} column giving the
index of each row. Otherwise, \code{rowid} will be added with a warning.}

\item{order}{String giving the name or index of the list elements in the
order they should be row-bind, with elements on the left winning over
elements on the right.}
}
\value{
A dataframe.
}
\description{
This function orders the output of \code{\link[=get_equations]{get_equations()}} by equation type, then
reduces it in order, by overwriting equations of each type with the
equations of a higher-priority type.
}
\examples{
allodb::scbi_tree1 \%>\%
  census_species(allodb::scbi_species, "scbi") \%>\%
  get_equations() \%>\%
  pick_best_equations()
}
\seealso{
\code{\link[=rowbind_inorder]{rowbind_inorder()}}.
}
