% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_bad_equations.R
\name{pick_equations}
\alias{pick_equations}
\title{Pick equations from \strong{allodb}.}
\usage{
pick_equations(site_species, specificity)
}
\arguments{
\item{site_species}{Dataframe. The species table of a ForestGEO site.}

\item{specificity}{String to match values from
\code{allodb::equations$allometry_specificity}.}
}
\value{
A dataframe.
}
\description{
From all equations in \strong{allodb}, pick equations for a given site and
specificity.
}
\examples{
pick_equations(allodb::scbi_species, "species")
}
