% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_allodb_master.R
\name{read_csv_as_chr}
\alias{read_csv_as_chr}
\title{Read a .csv file interpreting all columns as characters.}
\usage{
read_csv_as_chr(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path).}
}
\value{
A tibble.
}
\description{
This is a conservative way of reading data in a way that preserves all its
information.
}
\examples{
read_csv_as_chr("x\\n1")
}
