% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{references_metadata}
\alias{references_metadata}
\title{Table of metadata associated to the table \code{references}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.}
\usage{
references_metadata
}
\description{
Table of metadata associated to the table \code{references}.
}
\seealso{
Other database datasets: \code{\link{equations}},
  \code{\link{missing_values_metadata}},
  \code{\link{sites_info}}, \code{\link{sitespecies}},
  \code{\link{wsg}}
}
\concept{database datasets}
\keyword{datasets}
