% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_agb.R
\name{resample_agb}
\alias{resample_agb}
\title{Resample allodb equations to calibrate new allometries.}
\usage{
resample_agb(
  genus,
  coords,
  species = NULL,
  new_eqtable = NULL,
  wna = 0.1,
  w95 = 500,
  nres = 10000
)
}
\arguments{
\item{genus}{a character value, containing the genus (e.g. "Quercus") of the
tree.}

\item{coords}{a numeric vector of length 2 with longitude and latitude.}

\item{species}{a character value, containing the species (e.g. "rubra") of
the tree. Default is NULL, when no species identification is available.}

\item{new_eqtable}{Optional. An equation table created with the
add_equation() function. Default is the original allodb equation table.}

\item{wna}{a numeric vector, this parameter is used in the weight_allom function to determine
the dbh-related and sample-size related weights attributed to equations
without a specified dbh range or sample size, respectively. Default is 0.1}

\item{w95}{a numeric vector, this parameter is used in the weight_allom function to determine
the value at which the sample-size-related weight reaches 95\% of its
maximum value (max=1). Default is 500.}

\item{nres}{number of resampled values. Default is 1e4.}
}
\value{
A data frame (\code{tibble::tibble()} object) of resampled DBHs and associated AGB from the equation
table; the number of  resampled DBHs is proportional to the weight provided
by the \code{weight_allom} function.
}
\description{
After attributing a weight to each equation in allodb using the weight_allom
function, equations are then resampled within their original DBH range using
\code{resample_agb}: the number of resampled values for each equation is
proportional to its weight. It creates S3 objects of class "numeric".
}
\examples{
resample_agb(
  genus = "Quercus",
  species = "rubra",
  coords = c(-78.2, 38.9)
)

}
