% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scbi}
\alias{scbi}
\alias{scbi_tree1}
\alias{scbi_tree2}
\alias{scbi_species}
\alias{scbi_stem1}
\alias{scbi_stem2}
\title{Data from SCBI.}
\format{Objects of class tbl_df (inherit from tbl, data.frame). See structure
in \strong{Examples}.}
\source{
http://bit.ly/scbi-forestgeo-data
}
\usage{
scbi_tree1

scbi_tree2

scbi_species

scbi_stem1

scbi_stem2
}
\description{
Data from SCBI.
}
\examples{
dim(scbi_tree1)
dim(scbi_tree2)
str(scbi_tree2)

dim(scbi_stem1)
dim(scbi_stem2)
str(scbi_stem2)

str(scbi_species)
}
\concept{census datasets}
\keyword{datasets}
