% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scbi_stem1}
\alias{scbi_stem1}
\title{SCBI plot stem data (census 1)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55295 rows and 20 columns.}
\usage{
scbi_stem1
}
\description{
Dataset of stem data for census 1 at SCBI. Data is also public in Bourg et al. 2013.
}
\keyword{datasets}
