% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{site_eqn}
\alias{site_eqn}
\title{A dummy table of allometric equations by site in ForestGEO's network.}
\format{A data frame with 2 rows and 2 variables:
\itemize{
\item \code{site}: A character vector giving site names -- in lowercase.
\item \code{site_eqn}: A list of site-specific functions to calculate the biomass.
}}
\usage{
site_eqn
}
\description{
A dummy table of allometric equations by site in ForestGEO's network.
}
\examples{
library(dplyr)

site_eqn

# Pull the FAKE allometric equation of bci.
pull(filter(site_eqn, site == "bci"))
}
\keyword{datasets}
