% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{site_spp}
\alias{site_spp}
\title{A dummy table listing sites and species in ForestGEO's network.}
\format{A data frame with 12 rows and 2 variables:
\itemize{
\item \code{spp}: A character vector of species' names -- in lowercase.
\item \code{site}: A character vector giving site names -- in lowercase.
}}
\usage{
site_spp
}
\description{
This table relates the table of allometric equations by species with the
table of allometric equations by site.
}
\examples{
library(dplyr)

site_spp

full_join(
  full_join(site_spp, site_eqn),
  spp_eqn
)
}
\keyword{datasets}
