% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites_info}
\alias{sites_info}
\title{ForestGEO sites used in allodb}
\format{
A data frame with 24 observations and 6 variables:\preformatted{[,1]	Site	character
[,2]	site	character
[,3]	lat	decimal
[,4]	long	decimal
[,5]	elevation.m	numeric
[,6]	koppen	character

}
}
\usage{
sites_info
}
\description{
Table with geographical information for extratropical ForestGEO sites used in
allodb (n=24).
}
\details{
More details on geographical aspects of these ForestGEO sites can be found in
the accompanying manuscript.
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{missing_values}},
\code{\link{references}},
\code{\link{sitespecies}}
}
\concept{database datasets}
\keyword{datasets}
