% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_eqn.R
\name{table_eqn}
\alias{table_eqn}
\title{Table allometric equations.}
\usage{
table_eqn(x, eqn, sp, dbh, site = "my_site", eqn_type = c("site", "species",
  "stem"), source = c("user", "default"))
}
\arguments{
\item{x}{A dataframe, usually a ForestGEO-table.}

\item{eqn}{Name of the variable in \code{x} giving allometric equations.}

\item{sp}{Name of the variable in \code{x} species names.}

\item{dbh}{Name of the variable in \code{x} giving stem diameter.}

\item{site}{Character string giving the name of the site.}

\item{eqn_type}{One of stem, species or site.}

\item{source}{One of "user" or "default". Use "user" always, except
if you are a developer of \strong{allodb} and you are creating default tables
for ForestGEO.}

\item{stemid}{Name of the variable in \code{x} that uniquely identifies each stem.}
}
\value{
A dataframe.
}
\description{
Use these functions to create tables of allometric equations by site,
species, and stem.
}
\examples{
x <- tibble::tribble(~sp, ~equation, ~id, ~dbh,
                   "sp1",        NA, 200,   33,
                   "sp2", "3 * dbh", 111,   40,
                   "sp3", "7 * dbh", 221,   15,
                   "sp4", "5 * dbh", 332,   40)
x

table_eqn(x = x, eqn = "equation", sp = "sp", dbh = "dbh", site = "bci",
  eqn_type = "stem")

table_eqn(x = x, eqn = "equation", sp = "sp", dbh = "dbh", site = "bci",
  eqn_type = "species")

table_eqn(x = x, eqn = "equation", sp = "sp", dbh = "dbh", site = "bci",
  eqn_type = "site")
}
