% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_types.R
\name{type_allodb_master}
\alias{type_allodb_master}
\title{Help to read datasets safely, with consistent column type.}
\usage{
type_allodb_master()
}
\description{
For details see \code{\link[fgeo.tool:type_fgeo]{fgeo.tool::type_fgeo()}}.
}
\details{
Types' reference (see \code{\link[readr:read_csv]{readr::read_csv()}}.
\itemize{
\item c = character.
\item i = integer.
\item n = number.
\item d = double.
\item l = logical.
\item D = date.
\item T = date time.
\item t = time.
\item ? = guess.
\item _/- to skip the column.
}
}
