% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{user_eqn}
\alias{user_eqn}
\title{A fake data set of dbh an equations pretended to come from a user.}
\format{A tibble: 10 x 4:
\itemize{
\item \code{site} Site.
\item \code{spp} Species.
\item \code{dbh} Stem diameter.
\item \code{user_eqn} User's allometric-equation to calculate biomass by species.
}}
\usage{
user_eqn
}
\description{
This fake data set is made to contain some species for which the user
provides equations, some other species for which the user lacks equations,
and some rows with missing values of stem diameter (\code{dbh}).
}
\examples{
user_eqn
}
\keyword{datasets}
