% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_equation.R
\name{validate_equations}
\alias{validate_equations}
\title{Validate equations}
\usage{
validate_equations(equations, dbh = 1, h = 1, env = current_env())
}
\arguments{
\item{equations}{A character vector giving equations to validate.}

\item{dbh, h}{Independent variables}

\item{env}{An environment on which to evaluate the equations.}
}
\value{
This function is called for its side effect of throwing an error if
at least one equation is invalid. Valid equations are returned invisibly.
}
\description{
Validate equations
}
\examples{
# Good
validate_one_equation("dbh + 1")

# Fails
try(validate_one_equation("bad + 1"))

equations <- allodb::equations$equation_allometry
try(validate_equations(equations))

# FIXME: Some equations use "DBH" in uppercase. Is this intentional?
grep("DBH", equations, value = TRUE)[[1]]  # But we also have uppercase
grep("dbh", equations, value = TRUE)[[1]]  # I expected only lowercase
valid <- tolower(equations)

try(validate_equations(valid))
}
