% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wsg}
\alias{wsg}
\alias{wsg_metadata}
\title{Tables of wood density and associated metadata.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 549 rows and 8 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 8 columns.
}
\usage{
wsg

wsg_metadata
}
\description{
\itemize{
\item \code{wsg}: Table of allometric equations.
\item \code{wsg_metadata}: Table of metadata of \code{wsg}.
}
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{genus_family}},
\code{\link{missing_values_metadata}},
\code{\link{references_metadata}},
\code{\link{sites_info}},
\code{\link{sitespecies}}
}
\concept{database datasets}
\keyword{datasets}
