% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wsg_ctfs3}
\alias{wsg_ctfs3}
\title{Wood density data set from CTFS sites.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16558 rows and 9 columns.}
\source{
Suzanne (Lao LAOZ@si.edu).
}
\usage{
wsg_ctfs3
}
\description{
Wood density data set from CTFS sites.
}
\seealso{
Other data from tropical forests: \code{\link{WSG}},
  \code{\link{ficus}}, \code{\link{site.info}}
}
\keyword{datasets}
