% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_equations.R
\name{add_equation}
\alias{add_equation}
\title{Add new equations to the equation table (to then use in get_biomass function)}
\usage{
add_equation(
  taxa,
  allometry,
  coords,
  minDBH,
  maxDBH,
  sampleSize,
  unitDBH = "cm",
  unitOutput = "kg",
  inputVar = "DBH",
  outputVar = "Total aboveground biomass"
)
}
\arguments{
\item{taxa}{character string or vector specifying the taxon (or taxa) for which the allometry has been calibrated}

\item{allometry}{a character string with the allometric equation}

\item{coords}{a vector or matrix of coordinates (longitude, latitude) of the calibration data}

\item{minDBH}{numerical value, minimum DBH for which the equation is valid (in cm)}

\item{maxDBH}{numerical value, maximum DBH for which the equation is valid (in cm)}

\item{sampleSize}{number of measurements with which the allometry was calibrated}

\item{unitDBH}{character string with unit of DBH in the equation (either \code{cm}, \code{mm} or \code{inch}). Default is \code{cm}.}

\item{unitOutput}{character string with unit of equation output (either \code{g}, \code{kg}, \code{Mg} or \code{lbs} if the output is a mass, or \code{m} if the output is a height).}

\item{inputVar}{independent variable(s) needed in the allometry. Default is \code{DBH}, other option is \verb{DBH, H}.}

\item{outputVar}{dependent variable estimated by the allometry. Default is \verb{Total aboveground biomass}.}
}
\value{
A new equation table including all default equations and the additional equation (equation_id = \code{new})
}
\description{
This function creates S3 objects of class "numeric". Users can add multiple equations.
}
\examples{
new_equations <- add_equation(
  taxa = "Faga",
  allometry = "exp(-2+log(dbh)*2.5)",
  coords = c(-0.07, 46.11),
  minDBH = 5,
  maxDBH = 50,
  sampleSize = 50
)
}
