% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allodb-package.R
\docType{package}
\name{allodb-package}
\alias{allodb}
\alias{allodb-package}
\title{allodb: Tree Biomass Estimation at Extra-Tropical Forest Plots}
\description{
Standardize and simplify the tree biomass estimation process across globally distributed extratropical forests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/allodb/}
  \item \url{https://github.com/ropensci/allodb}
  \item Report bugs at \url{https://github.com/ropensci/allodb/issues}
}

}
\author{
\strong{Maintainer}: Erika Gonzalez-Akre \email{GonzalezEB@si.edu} (\href{https://orcid.org/0000-0001-8305-6672}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Camille Piponiot \email{camille.piponiot@gmail.com} (\href{https://orcid.org/0000-0002-3473-1982}{ORCID})
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})
  \item Kristina Anderson-Teixeira \email{TeixeiraK@si.edu} (\href{https://orcid.org/0000-0001-8461-9713}{ORCID})
}

Other contributors:
\itemize{
  \item Jeffrey Hanson \email{jeffrey.hanson@uqconnect.edu.au} [reviewer]
  \item Jonas Stillhard \email{jonas.stillhard@wsl.ch} [reviewer]
}

}
\keyword{internal}
