% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allodb-package.R
\docType{package}
\name{allodb-package}
\alias{allodb}
\alias{allodb-package}
\title{allodb: A Database of Allometric Equations for ForestGEO}
\description{
An R database for biomass estimation at globally distributed extratropical forest plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/allodb}
  \item Report bugs at \url{https://github.com/forestgeo/allodb/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

Authors:
\itemize{
  \item Erika Gonzalez-Akre \email{GonzalezEB@si.edu} (\href{https://orcid.org/0000-0001-8305-6672}{ORCID})
  \item Camille Piponiot \email{camille.piponiot@gmail.com} (\href{https://orcid.org/0000-0002-3473-1982}{ORCID})
  \item Kristina Anderson-Teixeira \email{TeixeiraK@si.edu}
}

}
\keyword{internal}
