% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allodb-package.R
\docType{package}
\name{allodb-package}
\alias{allodb}
\alias{allodb-package}
\title{allodb: An R Database For Tree Biomass Estimation At Extratropical Forest Plots}
\description{
Tool to standardize and simplify the tree biomass
    estimation process across globally distributed extratropical forests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/allodb}
  \item Report bugs at \url{https://github.com/forestgeo/allodb/issues}
}

}
\author{
\strong{Maintainer}: Erika Gonzalez-Akre \email{GonzalezEB@si.edu}

Authors:
\itemize{
  \item Camille Piponiot \email{camille.piponiot@gmail.com}
  \item Mauro Lepore \email{maurolepore@gmail.com}
  \item Kristina Anderson-Teixeira \email{TeixeiraK@si.edu}
}

}
\keyword{internal}
