% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allodb-package.R
\docType{package}
\name{allodb-package}
\alias{allodb}
\alias{allodb-package}
\title{allodb: A Database of Allometric Equations for ForestGEO}
\description{
A database of allometric equations for ForestGEO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/allodb}
  \item Report bugs at \url{https://github.com/forestgeo/allodb/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
