% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allodb-package.R
\docType{package}
\name{allodb-package}
\alias{allodb}
\alias{allodb-package}
\title{allodb: A Database of Allometric Equations for ForestGEO}
\description{
A database of allometric equations for ForestGEO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/allodb}
  \item Report bugs at \url{https://github.com/forestgeo/allodb/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Authors:
\itemize{
  \item Gonzalez-Akre Erika \email{GonzalezEB@@si.edu}
  \item Kristina Teixeira \email{TeixeiraK@@si.edu}
  \item Rutishauser Ervan \email{RutishauserE@@si.edu}
}

}
\keyword{internal}
