% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations}
\alias{equations}
\alias{equations_metadata}
\title{Tables of allometric equations and associated metadata for ForestGEO sites.}
\format{A data frame with 154 observations on 6 variables.\preformatted{[,1]	Ozone	 numeric	 Ozone (ppb)
[,2]	Solar.R	 numeric	 Solar R (lang)
[,3]	Wind	 numeric	 Wind (mph)
[,4]	Temp	 numeric	 Temperature (degrees F)
[,5]	Month	 numeric	 Month (1--12)
[,6]	Day	 numeric	 Day of month (1--31)
}}
\usage{
equations

equations_metadata
}
\description{
A dataset of best available allometry equations to calculate AGB per species
per ForestGEO site:
\itemize{
\item \code{equations}: Table of allometric equations.
\item \code{equations_metadata}: Table of metadata of \code{equations}.
}
}
\keyword{datasets}
