% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations}
\alias{equations}
\alias{equations_metadata}
\title{Table of allometric equations and associated metadata for ForestGEO sites.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 168 rows and 22 columns.}
\usage{
equations

equations_metadata
}
\description{
A compilation of best available allometry equations to calculate tree above-ground biomass (AGB)
per species at ForestGEO sites:
\itemize{
\item \code{equations}: Table of allometric equations.
\item \code{equations_metadata}: Metadata for \code{equations} table.
}
}
\keyword{datasets}
