% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations}
\alias{equations}
\alias{equations_metadata}
\title{Tables of allometric equations and associated metadata.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 421 rows and 14 columns.}
\usage{
equations

equations_metadata
}
\description{
\itemize{
\item \code{equations}: Table of allometric equations.
\item \code{equations_metadata}: Table of metadata of \code{equations}.
}
}
\keyword{datasets}
