% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations_metadata}
\alias{equations_metadata}
\title{Tables of allometric equations and associated metadata.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 47 rows and 7 columns.
}
\source{
See \link{references} for equations original sources.
}
\usage{
equations_metadata
}
\description{
\itemize{
\item \code{equations}: Table of allometric equations.
\item \code{equations_metadata}: Explanation of columns for \code{equations} table.
}
}
\details{
A compilation of best available allometry equations to calculate tree
above-ground biomass (AGB) per species based on extratropical ForestGEO
sites.
}
\examples{
#preview the datasets
print(head(equations))
print(head(equations_metadata))
}
\seealso{
Other database datasets: 
\code{\link{missing_values}},
\code{\link{references}},
\code{\link{sites_info}},
\code{\link{sitespecies}}
}
\concept{database datasets}
\keyword{datasets}
