% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equations}
\alias{equations}
\alias{equations_metadata}
\title{Tables of allometric equations and associated metadata for ForestGEO sites.}
\format{A data frame with 168 rows and 22 variables:\preformatted{# FIXME: Replace with real data.
[,1]	Ozone	 numeric	 Ozone (ppb)
[,2]	Solar.R	 numeric	 Solar R (lang)
[,3]	Wind	 numeric	 Wind (mph)
[,4]	Temp	 numeric	 Temperature (degrees F)
[,5]	Month	 numeric	 Month (1--12)
[,6]	Day	 numeric	 Day of month (1--31)
}}
\usage{
equations

equations_metadata
}
\description{
A compilation of best available allometry equations to calculate tree
above-ground biomass (AGB) per species at ForestGEO sites:
\itemize{
\item \code{equations}: Table of allometric equations.
\item \code{equations_metadata}: Table of metadata of \code{equations}.
}
}
\details{
\itemize{
\item equation_id: Unique equation identification number given arbitrarely.
\item equation_form: FIXME.
}
}
\seealso{
Other database datasets: \code{\link{missing_values_metadata}},
  \code{\link{references_metadata}},
  \code{\link{sites_info}}, \code{\link{sitespecies}},
  \code{\link{wsg}}
}
\concept{database datasets}
\keyword{datasets}
