% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_params.R
\name{est_params}
\alias{est_params}
\title{Function to calibrate new allometric equations from sampling previous ones}
\usage{
est_params(
  genus,
  species = NULL,
  coords,
  new_eqtable = NULL,
  wna = 0.1,
  w95 = 500
)
}
\arguments{
\item{genus}{A character vector, containing the genus (e.g. "Quercus") of
each tree.}

\item{species}{A character vector (same length as genus), containing the
species (e.g. "rubra")  of each tree. Default is NULL, when no
identification is available.}

\item{coords}{A numerical vector of length 2 with longitude and latitude (if
all trees were measured in the same location) or a matrix with 2 numerical
columns giving the coordinates of each tree.}

\item{new_eqtable}{Optional. An equation table created with the
add_equation() function. Default is the base allodb equation table.}

\item{wna}{this parameter is used in the weighting function to determine the
dbh-related and sample-size related weights attributed to equations without
a specified dbh range or sample size, respectively. Default is 0.1}

\item{w95}{this parameter is used in the weighting function to determine the
value at which the sample-size-related weight reaches 95\% of its maximum
value (max=1). Default is 500.}
}
\value{
A data frame of fitted coefficients (columns) of the log-log
regression: a (intercept), b (slope) and sigma (standard deviation). Each
row corresponds to a species x site combination. The back-transformed
equation is then AGB = exp(a) x DBH^b x + exp(0.5 x sigma^2).
}
\description{
This function creates S3 objects of class "numeric".
}
\examples{
data(scbi_stem1)
est_params(
  genus = scbi_stem1$genus,
  species = scbi_stem1$species,
  coords = c(-78.2, 38.9)
)

}
