% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genus_family}
\alias{genus_family}
\title{Genus and family table for selected ForestGEO sites}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 248 rows and 2 columns.
}
\usage{
genus_family
}
\description{
Genus and their associated family identified in the extratropical ForestGEO
sites used in allodb.
}
\examples{
genus_family
}
\seealso{
Other datasets: 
\code{\link{gymno_genus}},
\code{\link{koppenMatrix}},
\code{\link{scbi_stem1}},
\code{\link{shrub_species}}
}
\concept{datasets}
\keyword{datasets}
