% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass.R
\name{get_biomass}
\alias{get_biomass}
\title{Function to compute aboveground biomass (or other variables) from allometric
equations.}
\usage{
get_biomass(
  dbh,
  genus,
  species = NULL,
  coords,
  new_eqtable = NULL,
  wna = 0.1,
  w95 = 500
)
}
\arguments{
\item{dbh}{A numerical vector containing tree diameter at breast height (dbh)
measurements, in cm.}

\item{genus}{A character vector (same length as dbh), containing the genus
(e.g. "Quercus") of each tree.}

\item{species}{A character vector (same length as dbh), containing the
species (e.g. "rubra")  of each tree. Default is NULL, when no
identification is available.}

\item{coords}{A numerical vector of length 2 with longitude and latitude (if
all trees were measured in the same location) or a matrix with 2 numerical
columns giving the coordinates of each tree.}

\item{new_eqtable}{Optional. An equation table created with the
new_equations() function.}

\item{wna}{this parameter is used in the weighting function to determine the
dbh-related weight attributed to equations without a specified dbh range.
Default is 0.1}

\item{w95}{this parameter is used in the weighting function to determine the
value at which the sample-size-related weight reaches 95\% of its maximum
value (max=1). Default is 500.}

\item{wsteep}{this parameter controls the steepness of the dbh-related weight
in the weighting function. Default is 3.}
}
\value{
A vector of class "numeric" of the same length as dbh, containing AGB
value (in kg) for every stem, or the dependent variable as defined in
\code{var}.
}
\description{
This function creates S3 objects of class "numeric".
}
\details{
The function can run into some memory problems when used on large datasets
(usually several hundred thousand observations). In that case, see the 2nd
example below for how to apply the function to a splitted dataset.
}
\examples{
data(scbi_stem1)
get_biomass(
  dbh = scbi_stem1$dbh,
  genus = scbi_stem1$genus,
  species = scbi_stem1$species,
  coords = c(-78.2, 38.9)
)

# split dataset to avoid memory over usage
data_split <- split(scbi_stem1, cut(1:nrow(scbi_stem1), breaks = 10, labels = FALSE))
agb <- lapply(data_split, function(df) {
  get_biomass(
    dbh = df$dbh,
    genus = df$genus,
    species = df$species,
    coords = c(-78.2, 38.9)
  )
})
scbi_stem1$agb <- do.call(c, agb)
}
