% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass.R
\name{get_biomass}
\alias{get_biomass}
\title{Function to compute aboveground biomass (or other variables ) from allometric
equations.}
\usage{
get_biomass(
  dbh,
  h = NULL,
  genus = NULL,
  species = NULL,
  coords = NULL,
  var = "Total aboveground biomass",
  add_weight = FALSE
)
}
\arguments{
\item{dbh}{A numerical vector containing tree diameter at breast height (dbh)
measurements, in cm}

\item{h}{A numerical vector (same length as dbh) containing tree height
measurements, in m. Default is NULL, when no measurement is available.}

\item{genus}{A character vector (same length as dbh), containing the genus
(e.g. "Quercus") of each tree. Default is NULL, when no identification is
available.}

\item{species}{A character vector (same length as dbh), containing the
species (e.g. "rubra")  of each tree. Default is NULL, when no
identification is available.}

\item{coords}{A numerical vector of length 2 with longitude and latitude (if
all trees were measured in the same location) or a matrix with 2 numerical
columns giving the coordinates of each tree. Default is NULL when no
information is available.}

\item{var}{What dependent variable should be provided in the output? Default
is "Total aboveground biomass", other possible values are: "Bark biomass",
"Branches (dead)", "Branches (live)", "Branches total (live, dead)",
"Foliage total", "Height", "Leaves", "Stem (wood only)", "Stem biomass",
"Stem biomass (with bark)", "Stem biomass (without bark)", "Whole tree
(above and belowground)" , "Whole tree (above stump)". Be aware that only a
few equations exist for those other variables, so estimated values might
not be very acurate.}

\item{add_weight}{Should the relative weigth given to each equation in the
\code{equations} data frame be added to the output? Default is FALSE.}
}
\value{
A vector of class "numeric" of the same length as dbh, containing AGB
value (in kg) for every stem, or the dependent variable as defined in
\code{var}.
}
\description{
This function creates S3 objects of class "numeric".
}
\examples{
data(scbi_stem1)
get_biomass(
dbh = scbi_stem1$dbh,
genus = scbi_stem1$genus,
species = scbi_stem1$species,
coords = c(-78.2, 38.9)
)
}
