% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gymno_genus}
\alias{gymno_genus}
\title{Gymnosperms identified in selected ForestGEO sites}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 95 rows and 3 columns.
}
\usage{
gymno_genus
}
\description{
Table with genus and their associated family for Gymnosperms identified in
the ForestGEO sites used in allodb.
}
\examples{
gymno_genus
}
\seealso{
Other datasets: 
\code{\link{genus_family}},
\code{\link{koppenMatrix}},
\code{\link{scbi_stem1}},
\code{\link{shrub_species}}
}
\concept{datasets}
\keyword{datasets}
