% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koppenMatrix}
\alias{koppenMatrix}
\title{Koppen climate classification matrix}
\format{
A data frame with 900 observations and 3 variables:\preformatted{[,1]	zone1	character
[,2]	zone2	character
[,3]	wE	decimal

}
}
\usage{
koppenMatrix
}
\description{
A table built to facilitate the comparison between the Koppen climate of a site and equation in question. This #'table is used in the weighting scheme. More details can be found in the accompanying manuscript.
}
\details{
\itemize{
\item \code{koppenMatrix}: A dataframe with two columns depicting the 3-letter system of Köppen climate scheme and the given weight when comparing site/equation climate.
}
}
\keyword{datasets}
