% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{master}
\alias{master}
\title{Join allodb tables: \code{equations}, \code{sitespecies}, and \code{sites_info}.}
\usage{
master()
}
\value{
A dataframe.
}
\description{
Join allodb tables: \code{equations}, \code{sitespecies}, and \code{sites_info}.
}
\examples{
master()

\dontrun{
# Nice view in RStudio
View(master())
}
}
\seealso{
\code{\link[dplyr:full_join]{dplyr::full_join()}}.

Other functions to interact with the database: \code{\link{set_type}}
}
\concept{functions to interact with the database}
