% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{master}
\alias{master}
\title{Join the main allodb tables: \code{equations``, }sitespecies\code{, and }sites_info`.}
\usage{
master(.f = dplyr::full_join)
}
\arguments{
\item{.f}{A function, form the \strong{dplyr} package, used to join tables (e.g.
\code{dplyr::full_join()}).}
}
\value{
A dataframe.
}
\description{
Join the main allodb tables: \code{equations``, }sitespecies\code{, and }sites_info`.
}
\examples{
# Defaults to use `dplyr::full_join()`
master()

# But you can use other joins
master(.f = dplyr::left_join)
\dontrun{
# Nice view in RStudio
View(master())
}
}
\seealso{
\code{\link[dplyr:full_join]{dplyr::full_join()}}.
}
