% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_values}
\alias{missing_values}
\title{Explanations of missing values codes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
missing_values
}
\description{
Explanation of the codes used to indicate missing information in equation
table.
}
\examples{
missing_values
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{references}},
\code{\link{sites_info}},
\code{\link{sitespecies}}
}
\concept{database datasets}
\keyword{datasets}
