% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_values_metadata}
\alias{missing_values_metadata}
\title{Table of metadata associated to the table \code{missing_values}.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
missing_values_metadata
}
\description{
Example of alternative format for meta-data documentation:
\itemize{
\item NA: Not Applicable; Data does not apply to that particular case
}
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{references_metadata}},
\code{\link{sites_info}},
\code{\link{sitespecies}},
\code{\link{wsg}}
}
\concept{database datasets}
\keyword{datasets}
