% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_master.R
\name{read_master}
\alias{read_master}
\title{Help to read datasets safely.}
\usage{
read_master(file)
}
\description{
Reading the master table with this funcion ensures the following:
\itemize{
\item Missing values are consistent with \code{allodb::missing_values_metadata}
\item Columns have the type we expect, consistent with type_allodb_master().
}
}
\keyword{internal}
