% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_agb.R
\name{resample_agb}
\alias{resample_agb}
\title{Function to resample allodb equations to calibrate new allometries.}
\usage{
resample_agb(
  genus,
  species = NULL,
  coords,
  new_eqtable = NULL,
  wna = 0.1,
  w95 = 500
)
}
\arguments{
\item{genus}{A character value, containing the genus (e.g. "Quercus") of the
tree.}

\item{species}{A character value, containing the species (e.g. "rubra") of
the tree. Default is NULL, when no identification is available.}

\item{coords}{A numerical vector of length 2 with longitude and latitude.}

\item{new_eqtable}{Optional. An equation table created with the
add_equation() function. Default is the base allodb equation table.}

\item{wna}{this parameter is used in the weighting function to determine the
dbh-related and sample-size related weights attributed to equations without
a specified dbh range or sample size, respectively. Default is 0.1}

\item{w95}{this parameter is used in the weighting function to determine the
value at which the sample-size-related weight reaches 95\% of its maximum
value (max=1). Default is 500.}
}
\value{
A data frame of resampled DBHs and associated AGB from the equation
table; the number of  resampled DBHs is proportional to the weight provided
by the \code{weight_allom} function.
}
\description{
This function creates S3 objects of class "numeric".
}
\examples{
resample_agb(
  genus = "Quercus",
  species = "rubra",
  coords = c(-78.2, 38.9)
)

}
