% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scbi_stem1}
\alias{scbi_stem1}
\title{Tree census data from SCBI ForestGEO plot}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2287 rows and 6 columns.
}
\source{
Full data sets for tree census data at SCBI can be requested through
the ForestGEO portal (\url{https://forestgeo.si.edu/}). Census 1, 2, and 3 can
also be accessed at the public GitHub repository for SCBI-ForestGEO Data
(\url{https://github.com/SCBI-ForestGEO}).
}
\usage{
scbi_stem1
}
\description{
A table with tree data from the Smithsonian Conservation Biology Institute,
USA (SCBI) ForestGEO dynamics plot. These data is from the first tree census
in 2008, and only cover 1 hectare (SCBI is 25 ha).
}
\examples{
#preview the datasets
print(head(scbi_stem1))

}
\seealso{
Other datasets: 
\code{\link{genus_family}},
\code{\link{gymno_genus}},
\code{\link{koppenMatrix}},
\code{\link{shrub_species}}
}
\concept{datasets}
\keyword{datasets}
