% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scbi_stem1}
\alias{scbi_stem1}
\title{Tree census data from SCBI ForestGEO plot}
\format{
A data frame with 2287 observations and 6 variables:\preformatted{[,1]	treeID	numeric
[,2]	stemID	numeric
[,3]	dbh	numeric
[,4]	genus	character
[,5]	species	character
[,6]	Family	character

}
}
\source{
\itemize{
\item Full data sets for tree census data at SCBI can be requested through the ForestGEO portal (https://forestgeo.si.edu/). Census 1 and 2 can also be accessed at the public GitHub repository for SCBI-ForestGEO Data (https://github.com/SCBI-ForestGEO).
}
}
\usage{
scbi_stem1
}
\description{
A table with data from the Smithsonian Conservation Biology Institute, USA (SCBI) ForestGEO dynamics plot (1st tree census in 2008, 1 hectare).
}
\details{
\itemize{
\item \code{scbi_stem1}: A dataframe with tree data from the SCBI forest dynamic plot (1ha).
}
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{missing_values}},
\code{\link{references}},
\code{\link{sitespecies}}
}
\concept{database datasets}
\keyword{datasets}
