% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shrub_species}
\alias{shrub_species}
\title{Shrub species identified in selected ForestGEO sites}
\format{
An object of class \code{character} of length 1.
}
\usage{
shrub_species
}
\description{
Genus and species of shrubby plants identified in the 24 extratropical
ForestGEO sites used in allodb.
}
\examples{
#preview the dataset
print(head(shrub_species))

}
\seealso{
Other datasets: 
\code{\link{genus_family}},
\code{\link{gymno_genus}},
\code{\link{koppenMatrix}},
\code{\link{scbi_stem1}}
}
\concept{datasets}
\keyword{datasets}
