% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shrub_species}
\alias{shrub_species}
\title{List of shrub identified in selected ForestGEO sites}
\format{
A vector containing 179 observations
}
\usage{
shrub_species
}
\description{
A list with genus and species of shrubby plants identified in the ForestGEO sites used in allodb. The list is an input in the weight_allom() function.
}
\details{
\itemize{
\item \code{shrub_species}: A list with genus and species of shrubby plants identified in the extratropical ForestGEO #'sites used in allodb.
}
}
\keyword{datasets}
