% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites_info}
\alias{sites_info}
\title{ForestGEO sites used in allodb}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24 rows and 6 columns.
}
\usage{
sites_info
}
\description{
Table with geographical information for extratropical ForestGEO sites used in
allodb (n=24).
}
\details{
More details on geographical aspects of these ForestGEO sites can be found in
the accompanying manuscript.
}
\examples{
sites_info
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{missing_values}},
\code{\link{references}},
\code{\link{sitespecies}}
}
\concept{database datasets}
\keyword{datasets}
