% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitespecies}
\alias{sitespecies}
\alias{sitespecies_metadata}
\title{Tables of ForestGEO sites and tree species and associated metadata.}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 772 rows and 11 columns.}
\usage{
sitespecies

sitespecies_metadata
}
\description{
\itemize{
\item \code{sitespecies}: Table of ForestGEO sites and tree species.
\item \code{sitespecies_metadata}: Metadata for \code{sitespecies} table.
}
}
\seealso{
Other database datasets: \code{\link{equations}},
  \code{\link{missing_values_metadata}},
  \code{\link{references_metadata}},
  \code{\link{sites_info}}, \code{\link{wsg}}
}
\concept{database datasets}
\keyword{datasets}
