% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitespecies}
\alias{sitespecies}
\alias{sitespecies_metadata}
\title{Sites and tree species used in allodb and associated metadata}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1113 rows and 10 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
sitespecies

sitespecies_metadata
}
\description{
\itemize{
\item \link{sitespecies}: Table of extratropical ForestGEO sites in allodb (n=24) and
their tree species.
\item \link{sitespecies_metadata}: Metadata for \link{sitespecies} table.
}
}
\examples{
# preview the datasets
print(head(sitespecies))
print(head(sitespecies_metadata))
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{missing_values}},
\code{\link{references}},
\code{\link{sites_info}}
}
\concept{database datasets}
\keyword{datasets}
