% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitespecies}
\alias{sitespecies}
\alias{sitespecies}
\alias{sitespecies_metadata}
\title{Tables of sites and species and associated metadata.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 421 rows and 11 columns.}
\usage{
sitespecies

sitespecies_metadata
}
\description{
\itemize{
\item \code{sitespecies}: Table of sites and species.
\item \code{sitespecies_metadata}: Table of metadata of \code{sitespecies}.
}
}
\keyword{datasets}
