% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitespecies}
\alias{sitespecies}
\alias{sitespecies_metadata}
\title{Tables of sites and tree species used in allodb and associated metadata.}
\format{
A data frame with 1114 rows and 10 variables:\preformatted{[,1]	site	character
[,2]	family	character
[,3]	genus	character
[,4]	species	character
[,5]	variety	character
[,6]	subspecies	character
[,7]	latin_name	character
[,8]	species_code	character
[,9]	life_form	character
[,10]	notes	character

}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
sitespecies

sitespecies_metadata
}
\description{
\itemize{
\item \code{sitespecies}: Table of extratropical ForestGEO sites in allodb (n=24) and their tree species.
\item \code{sitespecies_metadata}: Metadata for \code{sitespecies} table.
}
}
\seealso{
Other database datasets: 
\code{\link{equations}},
\code{\link{missing_values}},
\code{\link{references}},
\code{\link{sites_info}}
}
\concept{database datasets}
\keyword{datasets}
