% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{spp_eqn}
\alias{spp_eqn}
\title{A dummy table of allometric equations by species in ForestGEO's network.}
\format{A data frame  with 8 rows and two variables:
\itemize{
\item \code{spp}: A character vector of species' names -- in lowercase.
\item \code{spp_eqn}: A list of species-specific functions to calculate the biomass.
}}
\usage{
spp_eqn
}
\description{
A dummy table of allometric equations by species in ForestGEO's network.
}
\examples{
spp_eqn
}
\keyword{datasets}
