% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_allodb_master.R
\name{type_allodb_master}
\alias{type_allodb_master}
\title{Help read allodb data safely.}
\usage{
type_allodb_master()
}
\description{
For details see \href{https://goo.gl/gVSMT8}{\code{fgeo.tool::type_fgeo()}}.
}
\details{
Types' reference (see \href{https://goo.gl/DtDvXq}{\code{readr::read_csv()}}.
\itemize{
\item c = character.
\item i = integer.
\item n = number.
\item d = double.
\item l = logical.
\item D = date.
\item T = date time.
\item t = time.
\item ? = guess.
\item _/- to skip the column.
}
}
\keyword{internal}
