% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_types.R
\name{type_allodb_master}
\alias{type_allodb_master}
\title{Help to read datasets safely, with consistent column type.}
\usage{
type_allodb_master()
}
\description{
For details see \href{https://goo.gl/gVSMT8}{fgeo.tool::type_fgeo()}.
}
\details{
Types' reference (see \href{https://goo.gl/DtDvXq}{readr::read_csv()}.
\itemize{
\item c = character.
\item i = integer.
\item n = number.
\item d = double.
\item l = logical.
\item D = date.
\item T = date time.
\item t = time.
\item ? = guess.
\item _/- to skip the column.
}
}
