% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_allom.R
\name{weight_allom}
\alias{weight_allom}
\title{Function to attribute weights for one species in one site.}
\usage{
weight_allom(
  genus,
  species = NULL,
  coords,
  new_eqtable = NULL,
  wna = 0.1,
  w95 = 500
)
}
\arguments{
\item{genus}{A character value, containing the genus (e.g. "Quercus") of the
tree.}

\item{species}{A character vector (same length as genus), containing the
species (e.g. "rubra") of the tree. Default is NULL, when no identification
is available.}

\item{coords}{A numerical vector of length 2 with longitude and latitude.}

\item{new_eqtable}{Optional. An equation table created with the
new_equations() function.}

\item{wna}{this parameter is used in the weighting function to determine the
sample-size related weights attributed to equations without a specified
sample size. Default is 0.1.}

\item{w95}{this parameter is used to determine the value at which the
sample-size-related weight reaches 95\% of its maximum value (max=1).
Default is 500.}
}
\value{
A vector with one weight for each equation.
}
\description{
This function creates S3 objects of class "numeric".
}
\examples{
weight_allom(
  genus = "Acer",
  species = "negundo",
  coords = c(-78.2, 38.9)
)

}
