% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WSG}
\alias{WSG}
\title{Wood density data set from CTFS sites.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16558 rows and 9 columns.}
\source{
Ervan Rutishauser (\link{er.rutishauser@gmail.com}), received on Sat,
Oct 28, 2017 at 4:05 AM.
}
\usage{
WSG
}
\description{
Wood density data set from CTFS sites.
}
\seealso{
\link{wsg_ctfs3}

Other data from tropical forests: \code{\link{ficus}},
  \code{\link{site.info}}, \code{\link{wsg_ctfs3}}
}
\keyword{datasets}
