% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wsg}
\alias{wsg}
\alias{wsg_metadata}
\title{Tables of wood density and associated metadata.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 377 rows and 8 columns.}
\usage{
wsg

wsg_metadata
}
\description{
\itemize{
\item \code{wsg}: Table of allometric equations.
\item \code{wsg_metadata}: Table of metadata of \code{wsg}.
}
}
\keyword{datasets}
